requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
python-dotenv==0.19.1
psycopg2-binary==2.9.3